requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.16.1
filelock==3.19.1
fsspec==2025.7.0
idna==3.10
inquirerpy==0.3.4
packaging==25.0
pfzy==0.3.4
prompt_toolkit==3.0.52
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.5
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.5.0
wcwidth==0.2.13
accelerate==1.10.1
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
Django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.16.1
filelock==3.19.1
fsspec==2025.7.0
hf-xet==1.1.9
idna==3.10
inquirerpy==0.3.4
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==2.3.2
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
packaging==25.0
pfzy==0.3.4
prompt_toolkit==3.0.52
psutil==7.0.0
python-dotenv==1.1.1
PyYAML==6.0.2
regex==2025.8.29
requests==2.32.5
safetensors==0.6.2
setuptools==80.9.0
sqlparse==0.5.3
sympy==1.14.0
tokenizers==0.22.0
torch==2.8.0
tqdm==4.67.1
transformers==4.56.0
triton==3.4.0
typing_extensions==4.15.0
urllib3==2.5.0
wcwidth==0.2.13
